Add DSL for new tests and some new tests
My first try to create some foundation for building better test suite
for the gem. The coverage of the current tests is pretty good, but the
problem with them - they are not atomic. Now, the test suite is just a
file with input lines, and a file with expected output. And if some
output line is not the same as the corresponding line of the expected
output, it will fail the test. The problem with that approach:

* It is hard to extend the tests
* It is hard to understand what we test exactly right now by that
  specific line
* It is pretty hard to understand what fails exactly
* They are very coupled (within one file), all next commands
  depend on previous commands
* If we need to add another output format (e.g., if we want to add XML
  or JSON to allow IDE to talk to 'debugger', we will need to duplicate
  the current test suite for that new output format

The good thing though - the tests are very decoupled from the
implementation of the debugger, they are more like 'Integration' tests.

We are in 2012 now, so let's try to use more modern approach for the
tests. I added some DSL for testing debugger, using MiniTest tests, and
the test suite for breakpoints as an example.