Add support of different Printers
So now we can specify different printers, which defines the text we send
to the interface. Added support of plain text printer (which displays
the texts currently used in debugger) and XML text printer, which I'm
going to use to recreate ruby-debug-ide functionality here, in the
'debugger' gem

You can specify what printer to use by Debugger.printer accessor.