Fix displaying of local variables
In XML, we now correctly display type and value (which was displayed
incorrectly, because we converted everything into a string by #inspect).

In Plain text, added exception to display nil as 'nil', because by
default it is displayed as '' by #to_s method. And use #to_s for
everything else